requirements.txt
python-telegram-bot==22.6
pandas==3.0.1
requests==2.32.5
openpyxl==3.1.5
python-dotenv==1.0.1
